Setup initial V2 programs boilerplate and scripts (#38)
- Adds vendored dependencies for all V2 program dependencies.
  - Includes bumping anchor to the latest version.
- Anchor 0.28.0 is only compatible with rust versions <= 0.78.0, but
anchor 0.31.0 requires >= 0.80.0. Because of this, I setup a `run.sh`
script that sets the correct versions for rust and anchor for each
building, cleaning, and running tests. We can add more functions to this
script later if needed.
- Adds Cargo.toml and Anchor.toml for v2 programs. 
- They have their own cargo and anchor config files because they will
have different dependency versions than V1, so this is the cleanest way
to define that separation.
- Adds initial boilierplate lib.rs for v2 programs.