Add V2 tests (#41)
- Copies over V1 tests to v2. 
- Makes a few small changes to tests to work with V2, not many changes
needed though, they're mostly just a copy of the existing tests.
- Create a new package.json in v2 directory because v2 contracts need
anchor v0.31.0
- V2 contract fixes to fix failing tests:
- Update `is_nonce_used`. The original change I made was failing to
deserialize the account correctly. The updated code passes tests for
both used and unused nonces.
- Update `token_pair` PDA seed to use `token_minter.bump` instead of
`token_messenger.bump`. This is because the `token_pair` bump is used in
other instructions where we don't have the `token_messenger` account.
- Also added `token_minter.bump` to the `token_pair` seed for the other
instructions where it was missing. This was revealed by running the
tests for v2 🙂
- Clean up readme
- Run V2 tests in Ci