Fix token_pair seed (#50)
In this PR:
https://github.com/circlefin/solana-cctp-contracts-private/pull/40#discussion_r2014514163
I added an unnecessary seed to the `token_pair` PDA because without it,
I was getting an error which we think was an Anchor IDL generation bug.
When trying to reproduce it, I discovered that as long as we call `to_
bytes()` in the function and pass the value as a reference (`&`) in the
seed, it works as expected.